refactor: contexts 폴더 제거
