+++ src/components/Providers/StudentInfoProvider/type.ts
@@ -1,0 +1,10 @@
+import { StudentGrade, StudentType } from '@/types/student';
+import { Merge } from '@/utils/type';
+
+export type UnfilledStudentInfoType = Merge<
+  StudentType,
+  {
+    grade: StudentGrade | undefined;
+    schoolId: number | undefined;
+  }
+>;
