--- src/contexts/CourseListContext.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-import { CourseType } from '@/schemas/courseSchema';
-
-export const CourseListContext = createContext<CourseType[]>([]);
