--- src/contexts/CourseTypeContext.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-import { CourseSelectionStepType } from '@/types/course';
-
-export const CourseTypeContext = createContext<CourseSelectionStepType>('MAJOR_REQUIRED');
